fix @xsi:type from bleeding into everything
@xsi:type is used in XSD's to define the type.  For most responses
from MantisConnect, an array of hash-items is returned, and calling
instance_variable_set on a type with @xsi:type breaks.  Since this isn't
used or needed on the client side, just remove it before mapping it to
IssueData or IssueHeaderData, or any of the other Data types that will
read and be used in making the web requests.