stub out Categories specs
I am working on trying to make sense of how Categories should work.

With the difficulty of working with the API as I have chosen (no Mocks,
lots of interaction with an external API), I am finally starting to feel
the pain of having to add setup and teardown of all the objects I'm sending
over the wire.  I'm sure I'll have to find a cleaner way to stub this out,
but I just don't know what that is, or if that will actually solve what
I'm trying to test for anymore.

Going to need to think on it a bit.