configure simplecov; pare down builder gem version
SimpleCov wasn't entirely configured right, so I changed a couple minor things, including
adding coverage/ to the .gitignore.

Builder is a dependency in Savon, but the range for version is pretty wide (>= 2.1.2), where
I have found JRuby does not work in 1.9 mode with builder 3.0.0, but works perfectly fine
with 2.1.2 (which is what it's base version is set to).  So, paring this down explicitly
should fix the JRuby issues