wrap specs in class; introduce 'let' block
After watching a video on Boston Ruby, it is better to wrap your files
in a class, so I've done that.  Nothing else changed in that instance.

Also, minitest 2.6.0 introduced the let syntax to mimic a behavior in RSpec,
so I made a change for @session to be a let :session block instead.  Not sure
about performance on it, but if I can get mocking to work with these tests, it
might not matter.