fill out more issue data
Finding that most of the Mantis features are just not implemented fully in
MantisConnect has led me to think that testing from here on out will probably
not matter much anyway, since I can't get a full coverage of the features in Mantis
as it is.

I'll probably regroup on this later, but right now I wnat to get the features I can
out there and accessible, since the API is regular enough that I can hope that
people would fumble around with it anyway (sorry for them, but I'd rather it work
if you know what you're after than not work at all).