rough-out rest of SOAP API
I have no tests for most of this, but since most of them are queries or
require very little in the way of manipulating data, it's pretty easy
to think that any issues I'll encounter will be because i misstyped something
or because MantisConnect doesn't work.  A third option is the user fat-fingered
or just didn't know what to put in to a field, which I can't help with, yet.